upgrade Symfony to 6.4 and fix deprecations
