+++ config/packages/doctrine.yaml
@@ -7,10 +7,11 @@ doctrine:
         #server_version: '16'
         use_savepoints: true
     orm:
+        controller_resolver:
+            auto_mapping: false
         auto_generate_proxy_classes: true
         enable_lazy_ghost_objects: true
         naming_strategy: doctrine.orm.naming_strategy.underscore_number_aware
-        auto_mapping: true
         mappings:
             App:
                 is_bundle: false
