+++ config/packages/framework.yaml
@@ -2,7 +2,9 @@
 framework:
     secret: '%env(APP_SECRET)%'
     csrf_protection: true
+    annotations: false
     http_method_override: false
+    handle_all_throwables: true
 
     # Enables session support. Note that the session will ONLY be started if you read or write from it.
     # Remove or comment this section to explicitly disable session support.
@@ -10,7 +12,6 @@ framework:
         handler_id: null
         cookie_secure: auto
         cookie_samesite: lax
-        storage_factory_id: session.storage.factory.native
 
     #esi: true
     #fragments: true
