--- config/packages/sensio_framework_extra.yaml
@@ -1,3 +1,0 @@
-sensio_framework_extra:
-    router:
-        annotations: false
