+++ config/packages/twig.yaml
@@ -4,6 +4,7 @@ twig:
     strict_variables: "%kernel.debug%"
     form_themes: ["bootstrap_3_layout.html.twig"]
     exception_controller: null
+    file_name_pattern: "*.twig"
 
 when@test:
     twig:
