+++ config/routes/security.yaml
@@ -1,0 +1,3 @@
+_security_logout:
+    resource: security.route_loader.logout
+    type: service
