+++ src/Controller/DefaultController.php
@@ -11,6 +11,6 @@ class DefaultController extends AbstractController
     #[Route(path: '/', name: 'homepage')]
     public function index(): Response
     {
-        return $this->renderForm('default/index.html.twig');
+        return $this->render('default/index.html.twig');
     }
 }
