+++ src/Controller/Task/TaskAddController.php
@@ -1,0 +1,39 @@
+<?php
+
+namespace App\Controller\Task;
+
+use App\Entity\Task;
+use App\Form\TaskType;
+use Doctrine\ORM\EntityManagerInterface;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+
+class TaskAddController extends AbstractController
+{
+    public function __construct(
+        private EntityManagerInterface $entityManager,
+    ) {
+    }
+
+    #[Route(path: '/tasks/create', name: 'task_create', methods: [Request::METHOD_GET, Request::METHOD_POST])]
+    public function create(Request $request): Response
+    {
+        $task = new Task();
+        $form = $this->createForm(TaskType::class, $task);
+
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            $this->entityManager->persist($task);
+            $this->entityManager->flush();
+
+            $this->addFlash('success', 'La tâche a été bien été ajoutée.');
+
+            return $this->redirectToRoute('task_list');
+        }
+
+        return $this->render('task/create.html.twig', ['form' => $form]);
+    }
+}
