+++ src/Controller/Task/TaskRemoveController.php
@@ -1,0 +1,29 @@
+<?php
+
+namespace App\Controller\Task;
+
+use App\Entity\Task;
+use Doctrine\ORM\EntityManagerInterface;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\RedirectResponse;
+
+class TaskRemoveController extends AbstractController
+{
+    public function __construct(
+        private readonly EntityManagerInterface $entityManager,
+    ) {
+    }
+
+    #[Route(path: '/tasks/{id}/delete', name: 'task_delete', methods: [Request::METHOD_GET])]
+    public function deleteTask(Task $task): RedirectResponse
+    {
+        $this->entityManager->remove($task);
+        $this->entityManager->flush();
+
+        $this->addFlash('success', 'La tâche a bien été supprimée.');
+
+        return $this->redirectToRoute('task_list');
+    }
+}
