+++ src/Controller/Task/TaskToggleController.php
@@ -1,0 +1,29 @@
+<?php
+
+namespace App\Controller\Task;
+
+use App\Entity\Task;
+use Doctrine\ORM\EntityManagerInterface;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\RedirectResponse;
+
+class TaskToggleController extends AbstractController
+{
+    public function __construct(
+        private readonly EntityManagerInterface $entityManager,
+    ) {
+    }
+
+    #[Route(path: '/tasks/{id}/toggle', name: 'task_toggle', methods: [Request::METHOD_GET])]
+    public function toggleTask(Task $task): RedirectResponse
+    {
+        $task->toggle(!$task->isDone());
+        $this->entityManager->flush();
+
+        $this->addFlash('success', sprintf('La tâche %s a bien été marquée comme faite.', $task->getTitle()));
+
+        return $this->redirectToRoute('task_list');
+    }
+}
