--- src/Controller/TaskController.php
@@ -1,94 +1,0 @@
-<?php
-
-namespace App\Controller;
-
-use App\Entity\Task;
-use App\Form\TaskType;
-use App\Repository\TaskRepository;
-use Doctrine\ORM\EntityManagerInterface;
-use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\Routing\Annotation\Route;
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\HttpFoundation\RedirectResponse;
-use Symfony\Bridge\Doctrine\Attribute\MapEntity;
-
-class TaskController extends AbstractController
-{
-    public function __construct(
-        private EntityManagerInterface $entityManager,
-        private TaskRepository $taskRepository
-    ) {
-        $this->entityManager = $entityManager;
-        $this->taskRepository = $taskRepository;
-    }
-
-    #[Route(path: '/tasks', name: 'task_list', methods: [Request::METHOD_GET])]
-    public function list(): Response
-    {
-        return $this->renderForm('task/list.html.twig', ['tasks' => $this->taskRepository->findAll()]);
-    }
-
-    #[Route(path: '/tasks/create', name: 'task_create', methods: [Request::METHOD_GET, Request::METHOD_POST])]
-    public function create(Request $request): Response
-    {
-        $task = new Task();
-        $form = $this->createForm(TaskType::class, $task);
-
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $this->entityManager->persist($task);
-            $this->entityManager->flush();
-
-            $this->addFlash('success', 'La tâche a été bien été ajoutée.');
-
-            return $this->redirectToRoute('task_list');
-        }
-
-        return $this->renderForm('task/create.html.twig', ['form' => $form]);
-    }
-
-    #[Route(path: '/tasks/{id}/edit', name: 'task_edit', methods: [Request::METHOD_PATCH])]
-    public function edit(Task $task, Request $request): Response
-    {
-        $form = $this->createForm(TaskType::class, $task);
-
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $this->entityManager->flush();
-
-            $this->addFlash('success', 'La tâche a bien été modifiée.');
-
-            return $this->redirectToRoute('task_list');
-        }
-
-        return $this->renderForm('task/edit.html.twig', [
-            'form' => $form,
-            'task' => $task,
-        ]);
-    }
-
-    #[Route(path: '/tasks/{id}/toggle', name: 'task_toggle', methods: [Request::METHOD_GET])]
-    public function toggleTask(Task $task): RedirectResponse
-    {
-        $task->toggle(!$task->isDone());
-        $this->entityManager->flush();
-
-        $this->addFlash('success', sprintf('La tâche %s a bien été marquée comme faite.', $task->getTitle()));
-
-        return $this->redirectToRoute('task_list');
-    }
-
-    #[Route(path: '/tasks/{id}/delete', name: 'task_delete', methods: [Request::METHOD_GET])]
-    public function deleteTask(Task $task): RedirectResponse
-    {
-        $this->entityManager->remove($task);
-        $this->entityManager->flush();
-
-        $this->addFlash('success', 'La tâche a bien été supprimée.');
-
-        return $this->redirectToRoute('task_list');
-    }
-}
