--- src/Controller/UserController.php
@@ -1,77 +1,0 @@
-<?php
-
-namespace App\Controller;
-
-use App\Entity\User;
-use App\Form\UserType;
-use App\Repository\UserRepository;
-use Doctrine\ORM\EntityManagerInterface;
-use Symfony\Component\Routing\Annotation\Route;
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\PasswordHasher\Hasher\UserPasswordHasherInterface;
-
-class UserController extends AbstractController
-{
-    public function __construct(
-        private UserPasswordHasherInterface $passwordHasher,
-        private EntityManagerInterface $entityManager,
-        private UserRepository $userRepository
-    ) {
-        $this->passwordHasher = $passwordHasher;
-        $this->entityManager = $entityManager;
-        $this->userRepository = $userRepository;
-    }
-
-    #[Route(path: '/users', name: 'user_list', methods: [Request::METHOD_GET])]
-    public function list(): Response
-    {
-        return $this->renderForm('user/list.html.twig', ['users' => $this->userRepository->findAll()]);
-    }
-
-    #[Route(path: '/users/create', name: 'user_create', methods: [Request::METHOD_GET, Request::METHOD_POST])]
-    public function create(Request $request): Response
-    {
-        $user = new User();
-        $form = $this->createForm(UserType::class, $user);
-
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $hashedPassword = $this->passwordHasher->hashPassword(
-                $user,
-                $form->get('password')->getData()
-            );
-            $this->userRepository->upgradePassword($user, $hashedPassword);
-
-            $this->addFlash('success', "L'utilisateur a bien été ajouté.");
-
-            return $this->redirectToRoute('user_list');
-        }
-
-        return $this->renderForm('user/create.html.twig', ['form' => $form]);
-    }
-
-    #[Route(path: '/users/{id}/edit', name: 'user_edit', methods: [Request::METHOD_GET, Request::METHOD_POST])]
-    public function edit(User $user, Request $request): Response
-    {
-        $form = $this->createForm(UserType::class, $user);
-
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $hashedPassword = $this->passwordHasher->hashPassword(
-                $user,
-                $form->get('password')->getData()
-            );
-            $this->userRepository->upgradePassword($user, $hashedPassword);
-
-            $this->addFlash('success', "L'utilisateur a bien été modifié");
-
-            return $this->redirectToRoute('user_list');
-        }
-
-        return $this->renderForm('user/edit.html.twig', ['form' => $form, 'user' => $user]);
-    }
-}
